pyproject.toml
[project]
name = "detector-backend"
version = "1.0.0"
description = "Logo Detection API with Django and MySQL"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.0",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.0",
    "opencv-python>=4.12.0.88",
    "ultralytics>=8.3.170",
    "pillow>=11.3.0",
    "numpy>=2.2.6",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "httpx>=0.27.0",
    "requests>=2.31.0",
    "drf-spectacular>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-django>=4.5.2",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
